refactor(studio): restructure components for agent management
