+++ apps/studio.giselles.ai/app/(main)/apps/components/create-agent-button.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+import { useFormStatus } from "react-dom";
+
+export function CreateAgentButton() {
+	const { pending } = useFormStatus();
+	return (
+		<Button type="submit" disabled={pending} data-loading={pending}>
+			New App +
+		</Button>
+	);
+}
