+++ apps/studio.giselles.ai/app/(main)/apps/myapps/page.tsx
@@ -6,7 +6,9 @@ import { formatTimestamp } from "@giselles-ai/lib/utils";
 import { and, desc, eq, isNotNull } from "drizzle-orm";
 import Link from "next/link";
 import { type ReactNode, Suspense } from "react";
-import { DeleteAgentButton, DuplicateAgentButton, Toasts } from "../components";
+import { DeleteAgentButton } from "../components/delete-agent-button";
+import { DuplicateAgentButton } from "../components/duplicate-agent-button";
+import { Toasts } from "../components/toasts";
 
 function DataList({ label, children }: { label: string; children: ReactNode }) {
 	return (
